Added the "title" query filter
